Implement Grid based NN Searcher
this implementation is based on implementing a row major two dimensional
array of IntArrayList's and mapping the coordinates into it.
This is as fast or faster as HashNN and takes less then 100 MB RAM.
It could further be improved for RAM usage by using an offset array
to get rid of the IntArrayList objects but this will decrease
code readability a lot.