Infrastructure to request with different protocols
this commit pulls the data of a ComputeRequest out into it's own
object which extends RequestData. We implement a ClassicRequestData
class that handles the standard protocol.
The RequestData is parsed by AlgorithmFactories so as to allow
for other protocols based on the type of Algorithm used.
ClassicRequestData is parsed by GraphAlgorithmFactory, but can
be overriden to enable GraphAlgorithms to use their own format.